pyproject.toml
[build-system]
requires = ["setuptools>=61.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pocket_extraction"
version = "0.1.3"
authors = [
    {name = "Hanker", email = "wu_2018@mail.ustc.edu.cn"}
]
description = "A package for extracting ligands and binding pockets from PDB files"
readme = "README.md"
requires-python = ">=3.7"
keywords = ["pocket extraction", "ligand extraction", "PDB processing"]
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent"
]
dependencies = [
    "biopython",
    "rdkit",
    "numpy"
]

[project.urls]
Homepage = "https://github.com/HankerWu/pocket_extraction"

[tool.setuptools.packages.find]
where = ["src"]  

[tool.setuptools]
include-package-data = true

[project.scripts]
extract_pocket = "pocket_extraction.extract_pocket:main"
extract_ligand = "pocket_extraction.extract_ligand:main"
extract_ligand_and_pocket = "pocket_extraction.extract_ligand_and_pocket:main"

